requirements.txt
torchtext==0.4.0 #0.8.0
transformers==4.4.2
numpy==1.19.2 #1.19.5
pytorch_ignite==0.4.7
ignite==1.1.0
#scikit_learn==1.0.2

#export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/nfs/stak/users/sundaraa/mionda3/envs/hw1/lib